requirements.txt
aiofile==3.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
anyio==4.7.0
attrs==24.3.0
caio==0.9.17
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.0
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
